refactor(member-card): move out of modals/ to components/member/
member-card is a display component, not a modal — its only callers
(member-application, now also settings overlay) embed it inline.
